backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
httpx==0.28.1
humanfriendly==10.0
idna==3.10
loguru==0.7.3
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
pyasn1==0.4.8
pydantic==2.11.3
pydantic_core==2.33.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uvicorn==0.34.0
